Fixed precision for all amount related columns

+++ app/graphql/types/payment_plan_type.rb
@@ -7,8 +7,8 @@ module Types
     field :status, String, null: true
     field :plan_type, String, null: true
     field :percentage, String, null: true
-    field :plot_balance, Integer, null: true
-    field :pending_balance, Integer, null: true
+    field :plot_balance, Float, null: true
+    field :pending_balance, Float, null: true
     field :land_parcel, Types::LandParcelType, null: false
     field :invoices, [Types::InvoiceType], null: false
     field :start_date, GraphQL::Types::ISO8601DateTime, null: true
